// Case_Study/task2/src/app/model/facility.ts
export interface Facility {
  id?: number;
  facilityType?: string;
  name?: string;
  area?: number;
  rentCost?: number;
  peopleMax?: number;
  rentalType?: string;
  standardRoom?: string;
  descriptionOtherConvenience?: string;
  poolArea?: number;
  numberOfFloors?: number;
  facilityFree?: string;
  image?: string;
}